final versions of new diffusion app with parallel deposition
